refactoring phase 1
refactoring phase 1 : une seule classe Comments.
Verification de la validitée des données dans les Controllers.

+++ Controller/ControllerAdminPage.php
@@ -1,6 +1,7 @@
 <?php
 namespace App\Back;
 
+use App\Comments;
 use App\Front\CommentsFront;
 use App\View;
 
@@ -10,17 +11,19 @@ class ControllerAdminPage {
 
     public function __construct()
     {
-        $this->comments = new CommentsBack();
+        $this->comments = new Comments();
     }
 
     public function adminPage()
     {
         $display = new View();
-        $display->createViewAdminPage($this->comments->getComments());
+        $display->createViewAdminPage($this->comments->getCommentsPerStatus());
     }
 
     public function deleteComment($idComment){
-        $this->comments->deleteComment($idComment);
+        if(is_numeric($idComment) AND $this->comments->existComment($idComment)) {
+            $this->comments->deleteComment($idComment);
+        }
         $this->adminPage();
     }
 }
